model-preference-app2/requirements.txt
fastapi==0.108.0
uvicorn[standard]==0.27.0
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
httpx==0.25.0
python-dotenv==1.0.0
openai==1.3.0
aiohttp==3.9.1
python-multipart==0.0.6
jinja2==3.1.6